A big commit that splits the TypeFactory in a part that is simply a factory for Types (and implements canonicalization a.k.a. maximal sharing), and a TypeStore which wraps declarations of types.
This changes the interfaces of some classes, especially the ones that deal with abstract data-types and constructors. The methods of these classes actually need a TypeStore as a parameter to figure out which constructors belong to which abstract data-type.

TypeStores can import each other such that adt's can be defined in a modular fashion. The use of TypeStore's is not needed for those who do not use the adt's.

The big benefit here is that TypeStore is a plain Java object that gets allocated with 'new', while TypeFactory is a singleton. The state of pdb.values can now be explicitly handled by the API user instead of the global TypeFactory maintaining all of it.